ADHOC: rename service command (service -> worker)
