+++ core/.air.conf
@@ -11,7 +11,7 @@ cmd = "go build -o ./bin/flagbased ./cmd/flagbased"
 # Binary file yields from `cmd`.
 bin = "bin/flagbased"
 # Customize binary.
-full_bin = "./bin/flagbased service start -v=true"
+full_bin = "./bin/flagbased worker start -v=true"
 # Watch these filename extensions.
 include_ext = ["go", "tpl", "tmpl", "html"]
 # Ignore these filename extensions or directories.
