+++ core/cmd/worker/worker_streamer.go
@@ -1,11 +1,11 @@
-package service
+package worker
 
 import (
 	"core/internal/pkg/cmdutil"
 	"core/internal/pkg/srvenv"
 )
 
-// StreamerConfig API service configuration
+// StreamerConfig API worker configuration
 type StreamerConfig struct {
 	Host         string
 	StreamerPort int
@@ -21,7 +21,7 @@ func StartStreamer(senv *srvenv.Env, cfg StreamerConfig) {
 		cmdutil.VerboseFlag, cfg.Verbose,
 	).Int(
 		StreamerPortFlag, cfg.StreamerPort,
-	).Msg("Starting Streamer Service")
+	).Msg("Starting Streamer Worker")
 
 	senv.Log.Warn().Msg("Streamer has not been implemented yet.")
 }
